+++ LoopStructural/interpolators/_finite_difference_interpolator.py
@@ -7,6 +7,8 @@ import numpy as np
 
 from LoopStructural.utils.helper import get_vectors
 from ._discrete_interpolator import DiscreteInterpolator
+from LoopStructural.interpolators import InterpolatorType
+
 from ._operator import Operator
 
 from LoopStructural.utils import getLogger
@@ -50,7 +52,7 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
 
         self.vol = 1.0  # grid.step_vector[0] * grid.step_vector[1] * \
         # grid.step_vector[2]
-        self.type = "FDI"
+        self.type = InterpolatorType.FINITE_DIFFERENCE
 
     def _setup_interpolator(self, **kwargs):
         """
@@ -130,6 +132,7 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
         self.add_vaue_constraints(self.interpolation_weights["cpw"])
         self.add_tangent_constraints(self.interpolation_weights["tpw"])
         self.add_interface_constraints(self.interpolation_weights["ipw"])
+        self.add_inequality_constraints()
 
     def copy(self):
         """
@@ -159,12 +162,11 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
             node_idx, inside = self.support.position_to_cell_corners(points[:, :3])
             # print(points[inside,:].shape)
 
-            gi = np.zeros(self.support.n_nodes)
+            gi = np.zeros(self.support.n_nodes, dtype=int)
             gi[:] = -1
-            gi[self.region] = np.arange(0, self.nx)
+            gi[self.region] = np.arange(0, self.nx, dtype=int)
             idc = np.zeros(node_idx.shape)
             idc[:] = -1
-
             idc[inside, :] = gi[node_idx[inside, :]]
             inside = np.logical_and(~np.any(idc == -1, axis=1), inside)
             a = self.support.position_to_dof_coefs(points[inside, :3])
@@ -177,6 +179,40 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
                 w=w * points[inside, 4],
                 name="value",
             )
+            if np.sum(inside) <= 0:
+                logger.warning(
+                    f"{self.propertyname}: {np.sum(~inside)} value constraints not added: outside of model bounding box"
+                )
+
+    def add_inequality_constraints(self, w=1.0):
+        points = self.get_inequality_constraints()
+        # check that we have added some points
+        if points.shape[0] > 0:
+            node_idx, inside = self.support.position_to_cell_corners(points[:, :3])
+            # print(points[inside,:].shape)
+
+            gi = np.zeros(self.support.n_nodes, dtype=int)
+            gi[:] = -1
+            gi[self.region] = np.arange(0, self.nx, dtype=int)
+            idc = np.zeros(node_idx.shape, dtype=int)
+            idc[:] = -1
+
+            idc[inside, :] = gi[node_idx[inside, :]]
+            inside = np.logical_and(~np.any(idc == -1, axis=1), inside)
+            a = self.support.position_to_dof_coefs(points[inside, :3])
+            # a*=w
+            # a/=np.product(self.support.step_vector)
+            self.add_inequality_constraints_to_matrix(
+                a.T,
+                points[inside, 3],
+                points[inside, 4],
+                idc[inside, :],
+                name="value_inequality",
+            )
+            if np.sum(inside) <= 0:
+                logger.warning(
+                    f"{self.propertyname}: {np.sum(~inside)} value constraints not added: outside of model bounding box"
+                )
 
     def add_interface_constraints(
         self, w=1.0
@@ -242,41 +278,6 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
                     name="interface_{}".format(unique_id),
                 )
 
-        # if points.shape[0] > 1:
-        #     node_idx, inside = self.support.position_to_cell_corners(
-        #                     points[:, :3])
-        #     # print(points[inside,:].shape)
-
-        #     gi = np.zeros(self.support.n_nodes)
-        #     gi[:] = -1
-        #     gi[self.region] = np.arange(0, self.nx)
-        #     idc = np.zeros(node_idx.shape)
-        #     idc[:] = -1
-
-        #     idc[inside, :] = gi[node_idx[inside, :]]
-        #     inside = np.logical_and(~np.any(idc == -1, axis=1), inside)
-        #     a = self.support.position_to_dof_coefs(points[inside, :3]).T
-        #     # create oversided array for storing constraints
-        #     A = np.zeros((a.shape[0]*a.shape[0],a.shape[1]*2))
-        #     interface_idc = np.zeros((a.shape[0]*a.shape[0],a.shape[1]*2),dtype=int)
-        #     interface_idc[:] = -1
-        #     c_i = 0
-
-        #     for i in np.unique(points[np.logical_and(~np.isnan(points[:,3]),inside),3]):
-        #         mask = points[inside,3] == i
-        #         for p1 in range(points[inside][mask].shape[0]):
-        #             for p2 in range(p1+1,points[inside][mask].shape[0]):
-        #                 A[c_i,:8] = a[mask][p1,:]
-        #                 A[c_i,8:] -= a[mask][p2,:]
-        #                 interface_idc[c_i,:8] = idc[inside,:][mask,:][p1,:]
-        #                 interface_idc[c_i,8:] = idc[inside,:][mask,:][p2,:]
-        #                 c_i+=1
-        #     outside = ~np.any(interface_idc == -1, axis=1)
-
-        #     self.add_constraints_to_least_squares(A[outside,:] * w,
-        #                                             np.zeros(A[outside,:].shape[0]),
-        #                                             interface_idc[outside, :], name='interface')
-
     def add_gradient_constraints(self, w=1.0):
         """
 
@@ -310,7 +311,7 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
                 vertices,
                 T,
                 elements,
-                inside,
+                inside_,
             ) = self.support.get_element_gradient_for_location(points[inside, :3])
             # normalise constraint vector and scale element matrix by this
             norm = np.linalg.norm(points[:, 3:6], axis=1)
@@ -327,6 +328,10 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
             self.add_constraints_to_least_squares(
                 A, B, idc[inside, :], w=w * self.vol, name="gradient"
             )
+            if np.sum(inside) <= 0:
+                logger.warning(
+                    f"{self.propertyname}: {np.sum(~inside)} norm constraints not added: outside of model bounding box"
+                )
 
     def add_norm_constraints(self, w=1.0):
         """
@@ -361,7 +366,7 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
                 vertices,
                 T,
                 elements,
-                inside,
+                inside_,
             ) = self.support.get_element_gradient_for_location(points[inside, :3])
             # T*=np.product(self.support.step_vector)
             # T/=self.support.step_vector[0]
@@ -387,6 +392,10 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
                 w=w * self.vol,
                 name="norm",
             )
+            if np.sum(inside) <= 0:
+                logger.warning(
+                    f"{self.propertyname}: {np.sum(~inside)} norm constraints not added: outside of model bounding box"
+                )
 
     def add_gradient_orthogonal_constraints(self, points, vector, w=1.0, B=0):
         """
@@ -427,7 +436,7 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
                 vertices,
                 T,
                 elements,
-                inside,
+                inside_,
             ) = self.support.get_element_gradient_for_location(points[inside, :3])
             T /= norm[:, None, None]
 
@@ -437,6 +446,10 @@ class FiniteDifferenceInterpolator(DiscreteInterpolator):
             self.add_constraints_to_least_squares(
                 A, B, idc[inside, :], w=w * self.vol, name="gradient orthogonal"
             )
+            if np.sum(inside) <= 0:
+                logger.warning(
+                    f"{self.propertyname}: {np.sum(~inside)} gradient constraints not added: outside of model bounding box"
+                )
 
     def add_regularisation(self, operator, w=0.1):
         """