+++ LoopStructural/interpolators/supports/_3d_structured_tetra.py
@@ -214,7 +214,108 @@ class TetMesh(BaseStructuredSupport):
         tetra_return[inside, :] = tetras[mask, :]
         return vertices_return, c_return, tetra_return, inside
 
-    
+    def get_constant_gradient(self, region, direction=None):
+        """
+        Get the constant gradient for the specified nodes
+
+        Parameters
+        ----------
+        region : np.array(dtype=bool)
+            mask of nodes to calculate cg for
+
+        Returns
+        -------
+
+        """
+        """
+        Add the constant gradient regularisation to the system
+
+        Parameters
+        ----------
+        w (double) - weighting of the cg parameter
+
+        Returns
+        -------
+
+        """
+        if direction is not None:
+            logger.info("Running constant gradient")
+            elements_gradients = self.get_element_gradients(np.arange(self.ntetra))
+            if elements_gradients.shape[0] != direction.shape[0]:
+                logger.error(
+                    "Cannot add directional CG, vector field is not the correct length"
+                )
+                return
+            region = region.astype("int64")
+
+            neighbours = self.get_neighbours()
+            elements = self.get_elements()
+            idc, c, ncons = fold_cg(
+                elements_gradients,
+                direction,
+                neighbours.astype("int64"),
+                elements.astype("int64"),
+                self.nodes,
+            )
+
+            idc = np.array(idc[:ncons, :])
+            c = np.array(c[:ncons, :])
+            B = np.zeros(c.shape[0])
+            return c, idc, B
+        if self.cg is None:
+            logger.info("Running constant gradient")
+            elements_gradients = self.get_element_gradients(np.arange(self.ntetra))
+            region = region.astype("int64")
+
+            neighbours = self.get_neighbours()
+            elements = self.get_elements()
+            idc, c, ncons = cg(
+                elements_gradients,
+                neighbours.astype("int64"),
+                elements.astype("int64"),
+                self.nodes,
+                region.astype("int64"),
+            )
+
+            idc = np.array(idc[:ncons, :])
+            c = np.array(c[:ncons, :])
+            B = np.zeros(c.shape[0])
+            self.cg = (c, idc, B)
+        return self.cg[0], self.cg[1], self.cg[2]
+
+    def get_constant_norm(self, region):
+        """
+        Get the constant gradient for the specified nodes
+
+        Parameters
+        ----------
+        region : np.array(dtype=bool)
+            mask of nodes to calculate cg for
+
+        Returns
+        -------
+
+        """
+
+        logger.info("Running constant gradient")
+        elements_gradients = self.get_element_gradients(np.arange(self.ntetra))
+        region = region.astype("int64")
+
+        neighbours = self.get_neighbours()
+        elements = self.get_elements()
+        idc, c, ncons = constant_norm(
+            elements_gradients,
+            neighbours.astype("int64"),
+            elements.astype("int64"),
+            self.nodes,
+            region.astype("int64"),
+        )
+
+        idc = np.array(idc[:ncons, :])
+        c = np.array(c[:ncons, :])
+        B = np.zeros(c.shape[0])
+
+        return c, idc, B
 
     def get_elements(self):
         """
@@ -349,11 +450,15 @@ class TetMesh(BaseStructuredSupport):
                 ],
             ]
         )
+        # m[~inside,:,:] = np.nan
         I = np.array(
             [[-1.0, 1.0, 0.0, 0.0], [-1.0, 0.0, 1.0, 0.0], [-1.0, 0.0, 0.0, 1.0]]
         )
         m = np.swapaxes(m, 0, 2)
-        element_gradients = np.linalg.inv(m)
+        element_gradients = np.zeros_like(m)
+        element_gradients[:] = np.nan
+        element_gradients[inside, :, :] = np.linalg.inv(m[inside, :, :])
+        # element_gradients = np.linalg.inv(m)
 
         element_gradients = element_gradients.swapaxes(1, 2)
         element_gradients = element_gradients @ I