+++ LoopStructural/modelling/core/geological_model.py
@@ -139,9 +139,7 @@ class GeologicalModel:
         # we want to rescale the model area so that the maximum length is
         # 1
         self.origin = np.array(origin).astype(float)
-        originstr = "Model origin: {} {} {}".format(
-            self.origin[0], self.origin[1], self.origin[2]
-        )
+        originstr = f"Model origin: {self.origin[0]} {self.origin[1]} {self.origin[2]}"
         logger.info(originstr)
         self.maximum = np.array(maximum).astype(float)
         maximumstr = "Model maximum: {} {} {}".format(
@@ -212,7 +210,7 @@ class GeologicalModel:
         use_thickness=True,
         vector_scale=1,
         gradient=False,
-        **kwargs
+        **kwargs,
     ):
         """Alternate constructor for a geological model using m2l output
 
@@ -232,6 +230,8 @@ class GeologicalModel:
         """
         from LoopStructural.modelling.input.map2loop_processor import Map2LoopProcessor
 
+        log_to_file(f"{m2l_directory}/loopstructural_log.txt")
+        logger.info("Creating model from m2l directory")
         processor = Map2LoopProcessor(m2l_directory, use_thickness)
         processor._gradient = gradient
         processor.vector_scale = vector_scale
@@ -263,20 +263,20 @@ class GeologicalModel:
 
     @classmethod
     def from_processor(cls, processor):
+        logger.info("Creating model from processor")
         model = GeologicalModel(processor.origin, processor.maximum)
         model.data = processor.data
         for i in processor.fault_network.faults:
+            logger.info(f"Adding fault {i}")
             model.create_and_add_fault(
                 i,
                 **processor.fault_properties.to_dict("index")[i],
-                faultfunction="BaseFault"
+                faultfunction="BaseFault",
             )
         for edge, properties in processor.fault_network.fault_edge_properties.items():
             if model[edge[1]] is None or model[edge[0]] is None:
                 logger.warning(
-                    "cannot add splay {} or {} are not in the model".format(
-                        edge[1], edge[0]
-                    )
+                    f"Cannot add splay {edge[1]} or {edge[0]} are not in the model"
                 )
                 continue
             splay = False
@@ -313,6 +313,7 @@ class GeologicalModel:
                 faults = None
                 if processor.fault_stratigraphy is not None:
                     faults = processor.fault_stratigraphy[s]
+                logger.info(f"Adding foliation {s}")
                 f = model.create_and_add_foliation(
                     s, **processor.foliation_properties[s], faults=faults
                 )
@@ -344,7 +345,7 @@ class GeologicalModel:
             logger.info("GeologicalModel initialised from file")
             return model
         else:
-            logger.error("{} does not contain a geological model".format(file))
+            logger.error(f"{file} does not contain a geological model")
             return None
 
     def __getitem__(self, feature_name):
@@ -422,7 +423,7 @@ class GeologicalModel:
             )
             return
         try:
-            logger.info("Writing GeologicalModel to: {}".format(file))
+            logger.info(f"Writing GeologicalModel to: {file}")
             pickle.dump(self, open(file, "wb"))
         except pickle.PicklingError:
             logger.error("Error saving file")
@@ -440,15 +441,14 @@ class GeologicalModel:
 
         if feature.name in self.feature_name_index:
             logger.info(
-                "Feature %s already exists at %i, overwriting"
-                % (feature.name, self.feature_name_index[feature.name])
+                f"Feature {feature.name} already exists at {self.feature_name_index[feature.name]}, overwriting"
             )
             self.features[self.feature_name_index[feature.name]] = feature
         else:
             self.features.append(feature)
             self.feature_name_index[feature.name] = len(self.features) - 1
             logger.info(
-                "Adding %s to model at location %i" % (feature.name, len(self.features))
+                f"Adding {feature.name} to model at location {len(self.features)}"
             )
         self._add_domain_fault_above(feature)
         self._add_unconformity_above(feature)
@@ -492,9 +492,7 @@ class GeologicalModel:
             except:
                 logger.error("Could not load pandas data frame from data")
                 raise BaseException("Cannot load data")
-        logger.info(
-            "Adding data to GeologicalModel with {} data points".format(len(data))
-        )
+        logger.info(f"Adding data to GeologicalModel with {len(data)} data points")
         self._data = data.copy()
         self._data["X"] -= self.origin[0]
         self._data["Y"] -= self.origin[1]
@@ -515,12 +513,19 @@ class GeologicalModel:
                     self._data[h] = 1.0
                 if h == "coord":
                     self._data[h] = 0
+                if h == "polarity":
+                    self._data[h] = 1.0
+        # LS wants polarity as -1 or 1, change 0 to -1
+        self._data.loc[self._data["polarity"] == 0, "polarity"] = -1.0
         self.data.loc[np.isnan(self.data["w"]), "w"] = 1.0
         if "strike" in self._data and "dip" in self._data:
             logger.info("Converting strike and dip to vectors")
             mask = np.all(~np.isnan(self._data.loc[:, ["strike", "dip"]]), axis=1)
-            self._data.loc[mask, gradient_vec_names()] = strike_dip_vector(
-                self._data.loc[mask, "strike"], self._data.loc[mask, "dip"]
+            self._data.loc[mask, gradient_vec_names()] = (
+                strike_dip_vector(
+                    self._data.loc[mask, "strike"], self._data.loc[mask, "dip"]
+                )
+                * self._data.loc[mask, "polarity"].to_numpy()[:, None]
             )
             self._data.drop(["strike", "dip"], axis=1, inplace=True)
 
@@ -625,7 +630,7 @@ class GeologicalModel:
         nelements=1e4,
         buffer=0.2,
         element_volume=None,
-        **kwargs
+        **kwargs,
     ):
         """
         Returns an interpolator given the arguments, also constructs a
@@ -672,15 +677,13 @@ class GeologicalModel:
                 for i in range(3):
                     if nsteps[i] < 3:
                         logger.error(
-                            "Number of steps in direction {} is too small, try increasing nelements".format(
-                                axis_labels[i]
-                            )
+                            f"Number of steps in direction {axis_labels[i]} is too small, try increasing nelements"
                         )
                 logger.error("Cannot create interpolator: number of steps is too small")
                 raise ValueError("Number of steps too small cannot create interpolator")
             # create a structured grid using the origin and number of steps
             if self.reuse_supports:
-                mesh_id = "mesh_{}".format(nelements)
+                mesh_id = f"mesh_{nelements}"
                 mesh = self.support.get(
                     mesh_id,
                     TetMesh(origin=bb[0, :], nsteps=nsteps, step_vector=step_vector),
@@ -735,9 +738,7 @@ class GeologicalModel:
                 for i in range(3):
                     if nsteps[i] < 3:
                         logger.error(
-                            "Number of steps in direction {} is too small, try increasing nelements".format(
-                                axis_labels[i]
-                            )
+                            f"Number of steps in direction {axis_labels[i]} is too small, try increasing nelements"
                         )
                 raise ValueError("Number of steps too small cannot create interpolator")
             # create a structured grid using the origin and number of steps
@@ -756,8 +757,8 @@ class GeologicalModel:
                     origin=bb[0, :], nsteps=nsteps, step_vector=step_vector
                 )
             logger.info(
-                "Creating regular grid with %i elements \n"
-                "for modelling using FDI" % grid.n_elements
+                f"Creating regular grid with {grid.n_elements} elements \n"
+                "for modelling using FDI"
             )
             return FDI(grid)
 
@@ -779,8 +780,8 @@ class GeologicalModel:
                     TetMesh(origin=bb[0, :], nsteps=nsteps, step_vector=step_vector),
                 )
             logger.info(
-                "Creating regular tetrahedron mesh with %i elements \n"
-                "for modelling using DFI" % mesh.ntetra
+                f"Creating regular tetrahedron mesh with {mesh.ntetra} elements \n"
+                "for modelling using DFI"
             )
             return DFI(mesh, kwargs["fold"])
         if interpolatortype == "Surfe" or interpolatortype == "surfe":
@@ -833,7 +834,7 @@ class GeologicalModel:
         # add data
         series_data = self.data[self.data["feature_name"] == series_surface_data]
         if series_data.shape[0] == 0:
-            logger.warning("No data for %s, skipping" % series_surface_data)
+            logger.warning("No data for {series_surface_data}, skipping")
             return
         series_builder.add_data_from_data_frame(series_data)
         self._add_faults(series_builder, features=faults)
@@ -874,7 +875,7 @@ class GeologicalModel:
         # add data
         series_data = self.data[self.data["feature_name"] == series_surface_data]
         if series_data.shape[0] == 0:
-            logger.warning("No data for %s, skipping" % series_surface_data)
+            logger.warning("No data for {series_surface_data}, skipping")
             return
         series_builder.add_data_from_data_frame(series_data)
         # self._add_faults(series_builder)
@@ -970,7 +971,7 @@ class GeologicalModel:
             logger.info("Using last feature as fold frame")
             fold_frame = self.features[-1]
         assert type(fold_frame) == FoldFrame, "Please specify a FoldFrame"
-        fold = FoldEvent(fold_frame, name="Fold_{}".format(foliation_data))
+        fold = FoldEvent(fold_frame, name=f"Fold_{foliation_data}")
         fold_interpolator = self.get_interpolator("DFI", fold=fold, **kwargs)
         if "fold_weights" not in kwargs:
             kwargs["fold_weights"] = {}
@@ -1036,7 +1037,7 @@ class GeologicalModel:
             logger.info("Using last feature as fold frame")
             fold_frame = self.features[-1]
         assert type(fold_frame) == FoldFrame, "Please specify a FoldFrame"
-        fold = FoldEvent(fold_frame, name="Fold_{}".format(fold_frame_data))
+        fold = FoldEvent(fold_frame, name=f"Fold_{fold_frame_data}")
         fold_interpolator = self.get_interpolator("DFI", fold=fold, **kwargs)
         gy_fold_interpolator = self.get_interpolator("DFI", fold=fold, **kwargs)
 
@@ -1051,7 +1052,7 @@ class GeologicalModel:
             name=fold_frame_data,
             fold=fold,
             frame=FoldFrame,
-            **kwargs
+            **kwargs,
         )
         fold_frame_builder.add_data_from_data_frame(
             self.data[self.data["feature_name"] == fold_frame_data]
@@ -1342,7 +1343,7 @@ class GeologicalModel:
         minor_axis=None,
         intermediate_axis=None,
         faultfunction="BaseFault",
-        **kwargs
+        **kwargs,
     ):
         """
         Parameters
@@ -1363,6 +1364,21 @@ class GeologicalModel:
         fault : FaultSegment
             created fault
         """
+        logger.info(f'Creating fault "{fault_surface_data}"')
+        logger.info(f"Displacement: {displacement}")
+        logger.info(f"Tolerance: {tol}")
+        logger.info(f"Fault function: {faultfunction}")
+        logger.info(f"Fault slip vector: {fault_slip_vector}")
+        logger.info(f"Fault center: {fault_center}")
+        logger.info(f"Major axis: {major_axis}")
+        logger.info(f"Minor axis: {minor_axis}")
+        logger.info(f"Intermediate axis: {intermediate_axis}")
+        fault_slip_vector = np.array(fault_slip_vector, dtype="float")
+        fault_center = np.array(fault_center, dtype="float")
+
+        for k, v in kwargs.items():
+            logger.info(f"{k}: {v}")
+
         if tol is None:
             tol = self.tol
         self.parameters["features"].append(
@@ -1373,6 +1389,9 @@ class GeologicalModel:
                 **kwargs,
             }
         )
+        if displacement == 0:
+            logger.warning(f"{fault_surface_data} displacement is 0")
+
         if "data_region" in kwargs:
             kwargs.pop("data_region")
             logger.error("kwarg data_region currently not supported, disabling")
@@ -1384,9 +1403,8 @@ class GeologicalModel:
             logger.error(
                 "Change interpolator to a discrete interpolation algorithm FDI/PLI"
             )
-            raise InterpolatorError(
-                "Faults not supported for {}".format(kwargs["interpolatortype"])
-            )
+            interpolatortype = kwargs["interpolatortype"]
+            raise InterpolatorError(f"Faults not supported for {interpolatortype}")
         fault_frame_builder = FaultBuilder(
             interpolator, name=fault_surface_data, model=self, **kwargs
         )
@@ -1394,12 +1412,29 @@ class GeologicalModel:
         fault_frame_data = self.data[
             self.data["feature_name"] == fault_surface_data
         ].copy()
+        trace_mask = np.logical_and(
+            fault_frame_data["coord"] == 0, fault_frame_data["val"] == 0
+        )
+        logger.info(f"There are {np.sum(trace_mask)} points on the fault trace")
+        if np.sum(trace_mask) == 0:
+            logger.error(
+                "You cannot model a fault without defining the location of the fault"
+            )
+            raise ValueError(f"There are no points on the fault trace")
+
         mask = np.logical_and(
             fault_frame_data["coord"] == 0, ~np.isnan(fault_frame_data["gz"])
         )
-        fault_normal_vector = (
-            fault_frame_data.loc[mask, ["gx", "gy", "gz"]].mean(axis=0).to_numpy()
+        vector_data = fault_frame_data.loc[mask, ["gx", "gy", "gz"]].to_numpy()
+        mask2 = np.logical_and(
+            fault_frame_data["coord"] == 0, ~np.isnan(fault_frame_data["nz"])
         )
+        vector_data = np.vstack(
+            [vector_data, fault_frame_data.loc[mask2, ["nx", "ny", "nz"]].to_numpy()]
+        )
+        fault_normal_vector = np.mean(vector_data, axis=0)
+        logger.info(f"Fault normal vector: {fault_normal_vector}")
+
         mask = np.logical_and(
             fault_frame_data["coord"] == 1, ~np.isnan(fault_frame_data["gz"])
         )
@@ -1427,15 +1462,17 @@ class GeologicalModel:
             logger.warning("Fault slip vector is nan, estimating from fault normal")
             strike_vector, dip_vector = get_vectors(fault_normal_vector[None, :])
             fault_slip_vector = dip_vector[:, 0]
-        if fault_center is not None:
+            logger.info(f"Estimated fault slip vector: {fault_slip_vector}")
+
+        if fault_center is not None and ~np.isnan(fault_center).any():
             fault_center = self.scale(fault_center, inplace=False)
-        if fault_center is None:
+        else:
             # if we haven't defined a fault centre take the center of mass for lines assocaited with
             # the fault trace
             if (
-                "centreEasting" in kwargs
-                and "centreNorthing" in kwargs
-                and "centreAltitude" in kwargs
+                ~np.isnan(kwargs.get("centreEasting", np.nan))
+                and ~np.isnan(kwargs.get("centreNorthing", np.nan))
+                and ~np.isnan(kwargs.get("centreAltitude", np.nan))
             ):
                 fault_center = self.scale(
                     np.array(
@@ -1469,17 +1506,15 @@ class GeologicalModel:
             minor_axis=minor_axis,
             major_axis=major_axis,
             intermediate_axis=intermediate_axis,
-            points=kwargs.get("points", None),
+            points=kwargs.get("points", True),
         )
 
         # if minor_axis == None or major_axis == None or intermediate_axis == None:
         #     fault_frame_builder.origin = self.bounding_box[0,:]
         #     fault_frame_builder.maximum = self.bounding_box[1,:]
         if "force_mesh_geometry" not in kwargs:
-            fault_frame_builder.set_mesh_geometry(
-                kwargs.get("fault_buffer", 0.2), 0
-            )  # ,
-            # np.rad2deg(np.arccos(np.dot(fault_normal_vector[:2],np.array([0,1])))))
+
+            fault_frame_builder.set_mesh_geometry(kwargs.get("fault_buffer", 0.2), 0)
         if "splay" in kwargs and "splayregion" in kwargs:
             fault_frame_builder.add_splay(kwargs["splay"], kwargs["splayregion"])
 
@@ -1491,7 +1526,7 @@ class GeologicalModel:
             fault_frame,
             displacement=displacement_scaled,
             faultfunction=faultfunction,
-            **kwargs
+            **kwargs,
         )
         fault.builder = fault_frame_builder
         for f in reversed(self.features):
@@ -1648,7 +1683,7 @@ class GeologicalModel:
                         )
                     ] = series["id"]
             if feature_id == -1:
-                logger.error("Model does not contain {}".format(group))
+                logger.error(f"Model does not contain {group}")
         return strat_id
 
     def evaluate_fault_displacements(self, points, scale=True):
@@ -1699,7 +1734,7 @@ class GeologicalModel:
         if feature_index > -1:
             return self.features[feature_index]
         else:
-            logger.error("{} does not exist!".format(feature_name))
+            logger.error(f"{feature_name} does not exist!")
             return None
 
     def evaluate_feature_value(self, feature_name, xyz, scale=True):
@@ -1789,10 +1824,7 @@ class GeologicalModel:
                 total_dof += f.interpolator.nx
         if verbose == True:
             print(
-                "Updating geological model. There are: \n"
-                "{} geological features that need to be interpolated\n".format(
-                    nfeatures
-                )
+                f"Updating geological model. There are: \n {nfeatures} geological features that need to be interpolated\n"
             )
 
         from tqdm.auto import tqdm
@@ -1805,8 +1837,8 @@ class GeologicalModel:
         with tqdm(total=nfeatures) as pbar:
             buf = 0
             for f in self.features:
-                pbar.set_description("Interpolating {}".format(f.name))
+                pbar.set_description(f"Interpolating {f.name}")
                 f.builder.up_to_date(callback=pbar.update)
 
         if verbose:
-            print("Model update took: {} seconds".format(time.time() - start))
+            print(f"Model update took: {time.time()-start} seconds")