+++ LoopStructural/modelling/fold/foldframe.py
@@ -156,8 +156,10 @@ class FoldFrame(StructuralFrame):
             ds = np.einsum("ij,ij->i", fold_axis, vv)
             flr = np.rad2deg(
                 np.arcsin(r2)
-            )  # np.where(np.abs(ds) < 0.5, np.rad2deg(np.arcsin(r2)),
-            # (- np.rad2deg(np.arcsin(r2))))
+            )  # np.where(ds > 0, np.rad2deg(np.arcsin(r2)),
+            # (- )))
+            flr[ds < 0] *= -1
+
             # flr = np.where(flr < -90, (180. + flr), flr)
             # flr = np.where(flr > 90, -(180. - flr), flr)
             return flr, s1